pyproject.toml
[tool.poetry]
name = "aes-cryptodrome"
version = "0.1.0"
description = ""
authors = ["Everson Esteves <119083241+esteveseverson@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pycryptodome = "^3.21.0"
inquirer = "^3.4.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.8.3"
taskipy = "^1.14.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
